[FEAT] 필요없는 파일 제거, 유진님 의견 반영
